refactor: Pagination 컴포넌트를 shared UI로 이동
- src/app/project/[project_name]/components/pagination/Pagination.tsx를
  src/shared/ui/InfinitePagination.tsx로 이동
- CollectClient에서 import 경로 업데이트
- 페이지 폴더의 재사용 가능한 컴포넌트를 shared로 이동하여
  프로젝트 구조 일관성 개선
- src/app/project/ 디렉토리 전체 삭제
